backend/requirements.txt
fastapi
uvicorn
pydantic
RPi.GPIO
